Reworked project for future expansion
GitHub:
- Added `CONTRIBUTING.md`
- Added Issue Templates
- Updated `README.md`

Code:
- Reworked file structure to be modular
- Reworked UI to support more components
- Various Fixes
- Styling Fixes
