+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,14 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: ''
+labels: enhancement
+assignees: ''
+
+---
+
+**Describe the suggestion**
+A clear and concise description of what you want to happen.
+
+**Any variations you've considered (optional)**
+A clear and concise description of any variations you've considered.
