+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+# Contributing to the project
+A compiled list of guidelines and information reguarding contributing to the project.
+## Development
+We invite anyone to contribute code via pull requests or file issues regarding problems or suggestions.
+
+### Setup:
+#### Prerequisites
+1. Node.js installed on your system (https://nodejs.org/en/)
+2. Yarn (npm install yarn)
+3. Forked Repository (Cloned if contributor)
+
+#### Running dev server
+1. Download dependencies with `yarn install` (executed in project root)
+2. Use `yarn run dev` to start the dev server
+3. Open dev server on `localhost:5000` in your browser
+
+### Submissions:
+**Issues** \
+Make sure to stick to issue templates.
+
+**Pull Requests** \
+All code submissions will be reviewed before being merged to main or dev.
+
+## Useful information
+All frameworks we use are rather beginner friendly and have amazing documentation.
+
+- Svelte ([Docs](https://svelte.dev/docs), [Examples](https://svelte.dev/examples))
+- TypeScript ([Docs](https://www.typescriptlang.org/docs/), 
+                [Docs for JS dev](https://www.typescriptlang.org/docs/handbook/typescript-in-5-minutes.html))
+- Sass ([Docs](https://sass-lang.com/documentation))
+- Bulma ([Docs](https://bulma.io/documentation/))
