+++ README.md
@@ -3,17 +3,10 @@ This is a calculator designed to provide the TToT for your ship.
 
 ## Links
 [GitHub Pages (Calculator itself)](https://th3-hero.github.io/AX-MTToT-Calculator/) \
-[Project board](https://github.com/Th3-Hero/AX-MTToT-Calculator/projects/1)
+[Project board](https://github.com/Th3-Hero/AX-MTToT-Calculator/projects)
 
-## Development:
-#### Prerequisites
-1. Node.js installed on your system (https://nodejs.org/en/)
-2. Repository cloned
-
-#### Running dev project
-1. Download dependencies with `npm install`
-2. Use `npm run dev` to start the dev server
-3. Open dev server on `localhost:5000`
+## Development
+See `CONTRIBUTING.md` for information on how you can help.
 
 ## Credits:
 Programing: [The Hero, Kuipers](https://bit.ly/3wzBDgD) \
