+++ src/components/inputs/ship-input/ShipInput.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
     import WeaponSelector from './weapon/Weapon.svelte';
     import DistroSelector from './distributor/Distro.svelte';
-    import { selectedWeapons } from '../../../typescript/store';
+    import { selectedWeapons } from '../../mttot/typescript/store';
 </script>
 
-<h2 class="mt-0 mb-2">Weapon Selection</h2>
+<h2 class="ml-2 mt-0 mb-2">Weapon Selection</h2>
 {#each Object.keys($selectedWeapons) as index}
     <WeaponSelector {index}/>
 {/each}
 
-<h2 class="mb-1">Distributor Selection</h2>
+<h2 class="ml-2 mb-1">Distributor Selection</h2>
 <DistroSelector/>
