+++ src/components/mttot/typescript/util.ts
