+++ src/typescript/Navbar.ts
@@ -1,0 +1,4 @@
+export enum NavbarTabs {
+    MTTOT = 'mttot',
+    GAUSS = 'gauss'
+}
