+++ src/typescript/data/dataFormat.ts
@@ -1,3 +1,7 @@
+// The below import is unused but necessary due to a Svelte issue with enums, described here:
+// https://www.reddit.com/r/sveltejs/comments/n87r2n/getting_a_typescript_enum_in_a_standalone_file_to/
+import { NavbarTabs } from '../Navbar'
+
 export enum WeaponType {
     FIXED = 'Fixed', GIMBAL = 'Gimballed', TURRET = 'Turret'
 }
